modified:   build/Desktop_Qt_6_8_2_MinGW_64_bit-Debug/Makefile.Debug 	modified:   build/Desktop_Qt_6_8_2_MinGW_64_bit-Debug/Makefile.Release 	modified:   build/Desktop_Qt_6_8_2_MinGW_64_bit-Debug/debug/data.o 	modified:   build/Desktop_Qt_6_8_2_MinGW_64_bit-Debug/debug/file.o 	modified:   build/Desktop_Qt_6_8_2_MinGW_64_bit-Debug/debug/find.o 	modified:   build/Desktop_Qt_6_8_2_MinGW_64_bit-Debug/debug/findWindow.o 	modified:   build/Desktop_Qt_6_8_2_MinGW_64_bit-Debug/debug/language.o 	modified:   build/Desktop_Qt_6_8_2_MinGW_64_bit-Debug/debug/mainwindow.o 	modified:   build/Desktop_Qt_6_8_2_MinGW_64_bit-Debug/debug/moc_find.o 	modified:   build/Desktop_Qt_6_8_2_MinGW_64_bit-Debug/debug/moc_mainwindow.o 	modified:   build/Desktop_Qt_6_8_2_MinGW_64_bit-Debug/debug/moc_table.o 	modified:   build/Desktop_Qt_6_8_2_MinGW_64_bit-Debug/debug/pr-products.exe 	modified:   build/Desktop_Qt_6_8_2_MinGW_64_bit-Debug/debug/qrc_translations.cpp 	modified:   build/Desktop_Qt_6_8_2_MinGW_64_bit-Debug/debug/qrc_translations.o 	modified:   build/Desktop_Qt_6_8_2_MinGW_64_bit-Debug/debug/row.o 	modified:   build/Desktop_Qt_6_8_2_MinGW_64_bit-Debug/debug/table.o 	modified:   build/Desktop_Qt_6_8_2_MinGW_64_bit-Debug/ui_find.h 	modified:   data.cpp 	modified:   data.hpp 	modified:   find.cpp 	modified:   find.h 	modified:   find.ui 	modified:   findWindow.cpp 	modified:   mainwindow.cpp 	modified:   mainwindow.h 	modified:   translations/my_en_US.ts 	modified:   translations/my_ru_RU.qm 	modified:   translations/my_ru_RU.ts

+++ data.hpp
@@ -49,8 +49,8 @@ public:
                 QString &name,
                 QString &type,
                 QString &mark,
-                double length,
-                double width,
+                int length,
+                int width,
                 double weight,
                 QString &fittings,
                 QString &date);
@@ -100,13 +100,13 @@ public:
      *   Метод, устанавливающий длину
      *   length длина
      */
-    void setLength(const double length);
+    void setLength(const int length);
 
     /**
      *   Метод, устанавливающий ширину
      *   width ширина
      */
-    void setWidth(const double width);
+    void setWidth(const int width);
 
 
     /**
@@ -141,10 +141,10 @@ public:
     QString getMark() const;
 
     ///   Метод, возвращающий длину
-    double getLenght() const;
+    int getLenght() const;
 
     ///   Метод, возвращающий ширину
-    double getWidth() const;
+    int getWidth() const;
 
     ///   Метод, возвращающий вес
     double getWeight() const;
